Stage3_Claude_Code_Engineering/Lesson06_Multimodal_RAG_System/GraphRAGAgent/backend/pyproject.toml
[project]
name = "graphrag-studio-backend"
version = "1.0.0"
description = "GraphRAG Studio — FastAPI backend service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "langextract[all]>=0.1.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "networkx>=3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
